Exclude Smoke Test run for certain branch (#6138)
* Exclude Smoke Test run for certain branch

This yml is hardcoded to 'translation-new-homepage' for now, but we may be able to devise a code for branches that need skipping in the future.

* Modified the code that gets the pr_details

using the issue number already have in the cat.

* Missing spaces in ${{ reference

* Made repo in the api call url dynamic

* Add BRANCHNAME to ./pr info

To be retrived by later workflows. e.g. smoketests.yml

* Using pr_information instead of API call

To get Branch Name but also modified the message I'm sending to Cypress Cloud slightly.